Tryna get filter trees back to their former glory but simplified. Stack merging algorithm is not working right
Former-commit-id: 540797eb391789e1773e822d6ae7a09fb2d9996c

+++ Anathema/Source/Tools/Filter/FilterTrees/IFilterTreeScanMVP.cs
@@ -26,7 +26,7 @@ namespace Anathema
         event FilterTreeScanEventHandler EventUpdateMemorySize;
 
         // Functions invoked by presenter (downstream)
-        void SetVariableSize(UInt64 VariableSize);
+        void SetVariableSize(Int32 VariableSize);
     }
 
     class FilterTreeScanPresenter : FilterPresenter
@@ -44,7 +44,7 @@ namespace Anathema
 
         #region Method definitions called by the view (downstream)
 
-        public void SetVariableSize(UInt64 VariableSize)
+        public void SetVariableSize(Int32 VariableSize)
         {
             if (VariableSize <= 0)
                 return;
